Too slow
This entire architecture appears to be slow. I don't know why in
particular. I don't have enough time to investigate further, hence I'll
change it to a block storage with specific client access library, and
implement the filesystem on top of it in the client. This will require
some synchronization between different clients through the block storage
(or an out-of-band lock manager, for example), but it will simplify the
backend a lot.